deps: bump nats.c to 3.9.3
Unfortunately, nats.c 3.10.0 does not build due to adding usage of the
type `SSL_verify_cb`, which is a "newer" OpenSSL API feature that
LibreSSL does not implement (LibreSSL's OpenSSL compatibility is
limited to the OpenSSL 1.0 API from whence it was forked, and the fact
that it worked with nats.c so far was, I think, largely a
coincidence). Unfortunately, the AYCB OpenSSL build wrapper is fairly
sloppy and does not build for any target besides x86_64 Linux. Since
I'm not actively using this library at this point, fixing the OpenSSL
build is a bridge too far for me. Maybe some good Samaritan can fix it
at some point and make it straightforward to swap over.